+++ backend/mora/auth/keycloak/uuid_extractor.py
@@ -1,26 +1,47 @@
 # SPDX-FileCopyrightText: 2018-2020 Magenta ApS
 # SPDX-License-Identifier: MPL-2.0
-
-import re
-from structlog import get_logger
+import asyncio
+import functools
 from typing import Set
 from typing import List
+from typing import Optional
+import re
+
+from more_itertools import one
+from structlog import get_logger
 from uuid import UUID
 
 from fastapi import Request
 
 from mora import common
 # import mora.main
+from mora.exceptions import ErrorCodes
+from mora.exceptions import HTTPException
 from mora.mapping import (
     ASSOCIATED_ORG_UNITS_FIELD,
+    ASSOCIATION,
     CHILDREN,
     DATA,
+    ENGAGEMENT,
+    EntityType,
+    MANAGER,
     ORG_UNIT,
     PARENT,
+    PERSON,
+    ROLE,
     TYPE,
+    USER_FIELD
 )
 
 logger = get_logger()
+
+
+def _get_terminate_entity_regex(entity: EntityType) -> re.Pattern:
+    return re.compile(
+        f'/service/{entity.value}/[0-9a-f]{{8}}-[0-9a-f]{{4}}-[0-9a-f]{{4}}-'
+        f'[0-9a-f]{{4}}-[0-9a-f]{{12}}/terminate'
+    )
+
 
 UUID_KEY = 'uuid'
 
@@ -34,19 +55,22 @@ UUID_KEY = 'uuid'
 
 # CREATE_OU = mora.main.app.url_path_for('create_org_unit')
 # TERMINATE_DETAIL = mora.main.app.url_path_for('terminate')
-# TERMINATE_UNIT = re.compile(
-#     mora.main.app.url_path_for(
-#         'terminate_org_unit',
-#         uuid='[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}'
-#     )
-# )
+# TERMINATE_UNIT = _get_terminate_entity_regex('terminate_org_unit')
+# TERMINATE_EMPLOYEE = _get_terminate_entity_regex('terminate_employee')
 
 CREATE_OU = '/service/ou/create'
 TERMINATE_DETAIL = '/service/details/terminate'
-TERMINATE_UNIT = re.compile(
-    '/service/ou/[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}'
-    '/terminate'
-)
+TERMINATE_UNIT = _get_terminate_entity_regex(EntityType.ORG_UNIT)
+TERMINATE_EMPLOYEE = _get_terminate_entity_regex(EntityType.EMPLOYEE)
+
+
+def return_value_logger(coro):
+    @functools.wraps(coro)
+    async def wrapper(*args, **kwargs):
+        result = await coro(*args, **kwargs)
+        logger.debug(f'{coro.__name__} returns', return_value=result)
+        return result
+    return wrapper
 
 
 def get_ancestor_uuids(tree: List[dict]) -> Set[UUID]:
@@ -54,7 +78,6 @@ def get_ancestor_uuids(tree: List[dict]) -> Set[UUID]:
     Recursively extract org unit UUIDs from org unit ancestor tree.
 
     :param tree: (sub)tree containing the descendants of an org unit
-    :param uuids: set of UUIDs to append the UUIDs of the descendants to
     :return: set of UUIDs of all the org units in the tree
     """
 
@@ -62,6 +85,7 @@ def get_ancestor_uuids(tree: List[dict]) -> Set[UUID]:
     # containing exactly one element, but let's leave this assertion until
     # the entire RBAC code is done
     assert len(tree) == 1  # Fix
+
     node = tree[0]
 
     uuids = {UUID(node.get(UUID_KEY))}  # Do not assume non-empty list
@@ -75,17 +99,21 @@ def get_ancestor_uuids(tree: List[dict]) -> Set[UUID]:
 # corresponding org units
 
 
-async def get_org_unit_uuids(request: Request) -> Set[UUID]:
+async def get_entity_uuids(request: Request) -> Set[UUID]:
     # State like pattern - choose appropriate UUID extraction strategy based
     # on the URL path of the incoming request
     """
-    Extract org unit UUIDs of the org units in the entire relevant branch of
+    Extract the UUID(s) for the relevant entity (org unit or employee).
+
+    For an org unit the coroutine gets the UUIDs in the entire relevant branch of
     the org unit ancestor tree. For example, if UnitA is a parent of UnitB
     which is a parent of UnitC, then will return a list containing the UUIDs
     of UnitA, UnitB and UnitC.
 
+    For an employee the coroutine just returns the UUID of the employee
+
     :param request: the incoming request to the endpoint
-    :return: list of the org unit UUIDs
+    :return: list of the entity UUID(s)
     """
 
     logger.debug('uuid_extract_strategy called')
@@ -93,6 +121,8 @@ async def get_org_unit_uuids(request: Request) -> Set[UUID]:
     if request.url.path == CREATE_OU:
         return await create_ou_extract_strategy(request)
     if TERMINATE_UNIT.match(request.url.path):
+        return path_extract_strategy(request)
+    if TERMINATE_EMPLOYEE.match(request.url.path):
         return path_extract_strategy(request)
     if request.url.path == TERMINATE_DETAIL:
         return await terminate_detail_extract_strategy(request)
@@ -160,11 +190,13 @@ async def terminate_detail_extract_strategy(request: Request) -> Set[UUID]:
     if payload[TYPE] == ORG_UNIT:
         return {UUID(payload[UUID_KEY])}
 
-    c = common.get_connector()
-    r = await c.organisationfunktion.get(uuid=payload[UUID_KEY])
-    org_unit_uuid = ASSOCIATED_ORG_UNITS_FIELD.get_uuid(r)
+    org_function = await _get_org_function(payload)
+    org_unit_uuid = ASSOCIATED_ORG_UNITS_FIELD.get_uuid(org_function)
+    if org_unit_uuid:
+        return {UUID(org_unit_uuid)}
+    employee_uuid = USER_FIELD.get_uuid(org_function)
 
-    return {UUID(org_unit_uuid)}
+    return {UUID(employee_uuid)}
 
 
 async def json_extract_strategy(request: Request) -> Set[UUID]:
@@ -182,31 +214,94 @@ async def json_extract_strategy(request: Request) -> Set[UUID]:
 
     logger.debug('json_uuid_extract_strategy called')
 
-    # payload is a list when the frontend calls /service/details/create
-    # and a dict when /service/details/edit is called
     payload = await request.json()
+    if isinstance(payload, dict):
+        payload = [payload]
 
-    if isinstance(payload, list):
-        # So far in the org unit RBAC case we have only experienced lists
-        # containing exactly one element, but let's leave this assertion until
-        # the entire RBAC code is done
-        assert len(payload) == 1
+    if not all(obj.get(TYPE) == payload[0].get(TYPE) for obj in payload):
+        logger.debug('Object types not identical')
+        raise HTTPException(
+            error_key=ErrorCodes.E_INVALID_INPUT,
+            message='Object types in payload list must be identical'
+        )
+    _type = payload[0].get(TYPE)
 
-        payload = payload[0]
+    async def obj_to_uuid(obj: dict) -> Optional[UUID]:
+        # All role, association and manager manipulations should be
+        # granted access via the org unit(s)
+        if _type in {ENGAGEMENT, ROLE, ASSOCIATION, MANAGER}:
+            if ORG_UNIT in obj:
+                return UUID(obj[ORG_UNIT][UUID_KEY])
+            if obj[TYPE] == ENGAGEMENT:
+                org_function = await _get_org_function(obj)
+                org_unit_uuid = ASSOCIATED_ORG_UNITS_FIELD.get_uuid(
+                    org_function
+                )
+                return UUID(org_unit_uuid)
+            return UUID(obj[DATA][ORG_UNIT][UUID_KEY])
+        # Creating or editing detail (address, IT-system,...)
+        if _type != ORG_UNIT:
+            if ORG_UNIT in obj:
+                return UUID(obj[ORG_UNIT][UUID_KEY])
+            return UUID(obj[PERSON][UUID_KEY])
+        return None
 
-    # Creating or editing org unit detail (address, IT-system,...)
-    if payload[TYPE] != ORG_UNIT:
-        return {UUID(payload[ORG_UNIT][UUID_KEY])}
+    uuids = set(await asyncio.gather(*(obj_to_uuid(obj) for obj in payload)))
+    uuids.discard(None)
+    if uuids:
+        return uuids
 
     # Renaming, moving or editing the details of the unit itself (i.e.
     # not address details and similar)
-    data = payload[DATA]
-    if PARENT in data and ORG_UNIT not in payload:
+
+    data = one(payload)[DATA]
+    if PARENT in data and ORG_UNIT not in one(payload):
         # Moving a unit
         return {UUID(data[UUID_KEY]), UUID(data[PARENT][UUID_KEY])}
 
     # Renaming or editing details on the unit itself
-    return {UUID(payload[DATA][UUID_KEY])}
+    return {UUID(one(payload)[DATA][UUID_KEY])}
+
+
+@return_value_logger
+async def get_entity_type(request: Request) -> EntityType:
+    if request.url.path == CREATE_OU or TERMINATE_UNIT.match(request.url.path):
+        return EntityType.ORG_UNIT
+
+    payload = await request.json()
+    if isinstance(payload, dict):
+        payload = [payload]
+
+    async def obj_to_type(obj: dict) -> EntityType:
+        if request.url.path == TERMINATE_DETAIL:
+            if obj[TYPE] == ORG_UNIT:
+                return EntityType.ORG_UNIT
+            else:
+                org_function = await _get_org_function(obj)
+                org_unit_uuid = ASSOCIATED_ORG_UNITS_FIELD.get_uuid(org_function)
+                return EntityType.ORG_UNIT if org_unit_uuid else EntityType.EMPLOYEE
+        if obj.get(TYPE) in {ENGAGEMENT, ROLE, ASSOCIATION, MANAGER}:
+            return EntityType.ORG_UNIT
+        if ORG_UNIT in obj or obj.get(TYPE) == ORG_UNIT:
+            return EntityType.ORG_UNIT
+
+        return EntityType.EMPLOYEE
+
+    types = await asyncio.gather(*(obj_to_type(obj) for obj in payload))
+    if not all(_type == types[0] for _type in types):
+        logger.debug('Types not identical')
+        raise HTTPException(
+            error_key=ErrorCodes.E_INVALID_INPUT,
+            message='Object types in payload list must be identical'
+        )
+
+    return types[0]
+
+
+async def _get_org_function(payload: dict) -> dict:
+    c = common.get_connector()
+    return await c.organisationfunktion.get(uuid=payload[UUID_KEY])
 
 # TODO: so far there are only integration tests covering this module -
-#  unit tests are still missing
+#  unit tests are still missing - and since this code probably will be
+#  removed later we should not spend too much effort on this